Split TestMain into two
runExternal uses `os.proc` which doesn't work on Scala Native
