+++ sjsonnet/test/src-jvm-native/sjsonnet/TestFiles.scala
@@ -1,0 +1,43 @@
+package sjsonnet
+
+object TestFiles {
+  val names = Seq(
+    "arith_bool",
+    "arith_float",
+    "arith_string",
+    "array",
+    "assert",
+    "binary",
+    "comments",
+    "condition",
+    "format",
+    "formatting_braces",
+    "formatting_braces2",
+    "functions",
+    "import",
+    "invariant",
+    "invariant_manifest",
+    "local",
+    "merge",
+    "null",
+    "object",
+    "oop",
+    "oop_extra",
+    "parsing_edge_cases",
+    "precedence",
+    "recursive_import_ok",
+    "recursive_object",
+    "sanity",
+    "sanity2",
+    "shebang",
+    "slice.sugar",
+    "std_all_hidden",
+    "text_block",
+    "unicode",
+    "unix_line_endings",
+    "unparse",
+    "verbatim_strings"
+    // disabled because it spams the console a lot
+//      "stdlib"
+  )
+}
